[Lab 08a] Ex - Refactored SpotMapsActivity.kt to MVP
