+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".activities.SpotMapsActivity"
+            android:name=".activities.SpotMapView"
             android:label="@string/title_spot_maps_activity" />
         <activity android:name=".views.spot.SpotView" />
         <activity android:name=".views.spotList.SpotListView">
