+++ app/src/main/java/com/archaeologicalfieldwork/activities/SpotMapPresenter.kt
@@ -1,0 +1,35 @@
+package com.archaeologicalfieldwork.activities
+
+
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.Marker
+import com.google.android.gms.maps.model.MarkerOptions
+import com.archaeologicalfieldwork.main.MainApp
+
+class SpotMapPresenter(val view: SpotMapView) {
+
+    var app: MainApp
+
+    init {
+        app = view.application as MainApp
+    }
+
+    fun doPopulateMap(map: GoogleMap) {
+        map.uiSettings.setZoomControlsEnabled(true)
+        map.setOnMarkerClickListener(view)
+        app.spots.findAll().forEach {
+            val loc = LatLng(it.lat, it.lng)
+            val options = MarkerOptions().title(it.title).position(loc)
+            map.addMarker(options).tag = it.id
+            map.moveCamera(CameraUpdateFactory.newLatLngZoom(loc, it.zoom))
+        }
+    }
+
+    fun doMarkerSelected(marker: Marker) {
+        val tag = marker.tag as Long
+        val spot = app.spots.findById(tag)
+        if (spot != null) view.showSpot(spot)
+    }
+}
