+++ app/src/main/java/com/archaeologicalfieldwork/activities/SpotMapView.kt
@@ -1,0 +1,64 @@
+package com.archaeologicalfieldwork.activities
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.Marker
+import kotlinx.android.synthetic.main.activity_spot_maps.*
+import com.archaeologicalfieldwork.R
+
+import com.archaeologicalfieldwork.helpers.readImageFromPath
+import com.archaeologicalfieldwork.models.SpotModel
+
+class SpotMapView : AppCompatActivity(), GoogleMap.OnMarkerClickListener {
+
+    lateinit var presenter: SpotMapPresenter
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_spot_maps)
+        setSupportActionBar(toolbar)
+        presenter = SpotMapPresenter(this)
+
+        mapView.onCreate(savedInstanceState);
+        mapView.getMapAsync {
+            presenter.doPopulateMap(it)
+        }
+    }
+
+    fun showSpot(spot: SpotModel) {
+        currentTitle.text = spot.title
+        currentDescription.text = spot.description
+        currentImage.setImageBitmap(readImageFromPath(this, spot.image))
+    }
+
+    override fun onMarkerClick(marker: Marker): Boolean {
+        presenter.doMarkerSelected(marker)
+        return true
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        mapView.onDestroy()
+    }
+
+    override fun onLowMemory() {
+        super.onLowMemory()
+        mapView.onLowMemory()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        mapView.onPause()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mapView.onResume()
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        mapView.onSaveInstanceState(outState)
+    }
+}
